Internet_Search_MCP/requirements.txt
pyyaml
ascii_colors
fastmcp
duckduckgo-search
pipmaster